Consolidated refresh functions to avoid duplicated code.
